// frontend/src/api/model/handlerDisasterResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 */
import type { HandlerPrefectureItem } from "./handlerPrefectureItem"

export interface HandlerDisasterResponse {
  affected_area_size?: number
  disaster_code?: string
  disaster_type?: string
  estimated_damage_amount?: number
  id?: string
  impact_level?: string
  name?: string
  occurred_at?: string
  prefecture?: HandlerPrefectureItem
  status?: string
  summary?: string
}